+++ pai-dlc-20201203/alibabacloud_pai_dlc20201203/models/_list_jobs_shrink_request.py
@@ -36,6 +36,7 @@ class ListJobsShrinkRequest(DaraModel):
         start_time: str = None,
         status: str = None,
         tags_shrink: str = None,
+        template_id: str = None,
         time_range_field: str = None,
         user_command_search: str = None,
         user_id_for_filter: str = None,
@@ -133,6 +134,7 @@ class ListJobsShrinkRequest(DaraModel):
         self.status = status
         # The tags.
         self.tags_shrink = tags_shrink
+        self.template_id = template_id
         self.time_range_field = time_range_field
         self.user_command_search = user_command_search
         # The user ID used to filter jobs.
@@ -236,6 +238,9 @@ class ListJobsShrinkRequest(DaraModel):
 
         if self.tags_shrink is not None:
             result['Tags'] = self.tags_shrink
+
+        if self.template_id is not None:
+            result['TemplateId'] = self.template_id
 
         if self.time_range_field is not None:
             result['TimeRangeField'] = self.time_range_field
@@ -342,6 +347,9 @@ class ListJobsShrinkRequest(DaraModel):
 
         if m.get('Tags') is not None:
             self.tags_shrink = m.get('Tags')
+
+        if m.get('TemplateId') is not None:
+            self.template_id = m.get('TemplateId')
 
         if m.get('TimeRangeField') is not None:
             self.time_range_field = m.get('TimeRangeField')